Added images, made bullets functional, modified fire function, cleaned code.

+++ Project X/Project X/PlayerShip.h
@@ -1,6 +1,5 @@
 #pragma once
 #include "Ship.h"
-#include <SDL.h>
 class PlayerShip :
 	public Ship
 {
@@ -9,6 +8,6 @@ public:
 	~PlayerShip();
 
 private:
-	Uint8 mNumLives;
+	int mNumLives;
 };
